Sync Mattlab with GeomagneticPhi sunspots.py improvements and updates
Summary of changes:
- **sunspots.py**: Updated with Astropy Time integration and retain HCSinclination use_orig parameter
  - Replaced deprecated helio_time imports with Astropy Time methods
  - Converted all time format conversions (MJD, fractional year, datetime)
  - Added helper functions: sph2cart(), cart2sph() for coordinate transformations
  - Updated all data loaders with Astropy equivalents while maintaining compatibility
  - Enhanced osf_and_streamer_belt_width() with improved flux conservation checks
  - Retained HCSinclination(yrs, min_yrs, use_orig=True) with legacy option for historical calculations

- **helio_time.py**: Fixed critical bug in doyyr2mjd() function
  - Added explicit int() casting in timedelta calculation to prevent floating-point errors
  - Ensures proper day-of-year conversion for large datasets

- **insitu.py**: Enhanced ICME list processing
  - Added support for new data quality flags ('... P', '... Q')
  - Improved NaN handling for edge cases

- **mplot.py**: Fixed visualization issues
  - Updated plotspeedmap() and plotbrmap() with proper meshgrid usage
  - Changed pcolor to pcolormesh for better rendering
  - Adjusted Bfield plot limits (-100 to 100) for better dynamic range

- **Data**: Updated Richardson-Cane ICME list
  - Added 26 new ICMEs covering 2024-2025 period
  - Updated existing 2024 entries with latest observations

- **New file**: code/dataloaders/get_SolarOrbiter_SWA_LL_data.py
  - Added Solar Orbiter low-latency data downloader for PAS and MAG instruments
  - Queries SOAR archive and downloads LL data products
  - Converts CDF to DataFrame with 1-hour resampling

All changes maintain backward compatibility with existing code while improving efficiency and data quality.

+++ code/helio_time.py
@@ -316,7 +316,7 @@ def doyyr2mjd(doy,yr):
     
     #create a datetime object at the start of the year, add doy as a delta
     def _create_dt_(doy,yr):
-        dt = datetime.datetime(yr,yr*0+1,yr*0+1) + datetime.timedelta(days = np.floor(doy)-yr*0+1)
+        dt = datetime.datetime(yr, yr*0+1, yr*0+1) + datetime.timedelta(days=int(np.floor(doy) - yr * 0 + 1))
         return dt
     vec_create_dt = np.vectorize(_create_dt_)
     dt = vec_create_dt(doy,yr)
